Eco benefits for multiple trees

This function calculates the benefits for an entire dataset, sometimes called a tree inventory. This works similar to i-Trees Eco software where the user supplies data, the common name field, the botanical name field, the dbh field, and the region. All calculations are done on unique species/dbh pairs to avoid redundant computation and speed up the calculation.

eco_run_all(data, common_col, botanical_col, dbh_col, region, n = 0.8,
  print_time = NULL)

Arguments

data
path to csv file containing tree inventory

common_col
the name of the column containing common names

botanical_col
the name of the column containing botanical names

dbh_col
the name of the column containing dbh values

region
region code, see species or benefits

n
guessing threshold from 0.0 to 1.0, defaults at 0.8

print_time
Logical TRUE or FALSE for printing the elapsed time